Gets rid of top-level 'five' namespace. Refs #4.
