+++ tox.ini
@@ -8,7 +8,7 @@ deps = -rrequirements-dev.txt
 commands =
     coverage erase
     coverage run -m pytest {posargs:tests}
-    coverage report --show-missing --fail-under 70
+    coverage report --show-missing --fail-under 85
     flake8 {[tox]project}.py tests setup.py
     pylint {[tox]project}.py tests setup.py
